+++ HISTORY.rst
@@ -21,6 +21,8 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* BSSFL-512: Add button to cancel MO no matter the state
+
 **Bugfixes**
 
 * BSSFL-527: Asset compute with move lines aggregation returns single move